Extract WorkloadIdentityFederationAuth to separate file
Move WorkloadIdentityFederationAuth and WorkloadIdentityFederationCredentials
classes to their own file for better code organization.
